Added some hierarchy and polymorphism

+++ src/include/Entidad.h
@@ -16,7 +16,7 @@ class Entidad
 {
 protected:
 
-    float health;
+    int health;
     sf::Vector2f size;
     float movementSpeed;
     sf::Texture texture;
